AOT Cache Compiler and Improvements to Cache Bootstrapper

+++ scripts/opt/collapseBundle.bash
@@ -1,3 +1,4 @@
+(set -o igncr) 2>/dev/null && set -o igncr; # For Cygwin on Windows compatibility
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
     target="phat-linux-x64/resources/app"
 fi
@@ -7,12 +8,14 @@ fi
 
 for f in $target/*.js
 do
-    printf "Collapsing bundle $f\n"
-    ./node_modules/.bin/bundle-collapser $f > tmp
-    if [ $? != 0 ]; then
-        rm temp
-    fi
-    if [ $? == 0 ]; then
-        mv tmp $f
+    if [[ "$f" != "$target/pileup.js" ]]; then
+        printf "Collapsing bundle $f\n"
+        ./node_modules/.bin/bundle-collapser $f > tmp
+        if [ $? != 0 ]; then
+            rm temp
+        fi
+        if [ $? == 0 ]; then
+            mv tmp $f
+        fi
     fi
 done
